Move Json logging into LoggerExtensions
